Move components to compose package and extract state package
