+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/compose/components/UserAvatarImage.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.main.jetpack.migration.components
+package org.wordpress.android.ui.main.jetpack.migration.compose.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
